Refactoring Mass Action classes, bug fixes

+++ app/Http/Utilities/Admin/Modules/Categories/PageCategoryEntity.php
@@ -4,7 +4,6 @@ namespace App\Http\Utilities\Admin\Modules\Categories;
 
 use App\Entities\PageCategory;
 use App\Interfaces\WebEntity;
-use Illuminate\Http\Request;
 use Auth;
 
 class PageCategoryEntity implements WebEntity
@@ -62,32 +61,11 @@ class PageCategoryEntity implements WebEntity
 
     public function destroy()
     {
-        Auth::user()->hasAccessOrRedirect('CATEGORY_DELETE');
+        if (!Auth::user()->hasAccess('CATEGORY_DELETE')) {
+            return redirect()->back()->with('error', 'You don\'t have rights to finish this action.');
+        }
+
         $this->item->delete();
-
-        return redirect(route('admin.pages.categories.index'));
-    }
-
-
-    public function apiUpdate(Request $request)
-    {
-        $access = Auth::user()->hasAccess('CATEGORY_EDIT');
-        if (!$access) return response()->json('No access');
-
-        $result = $this->item->update($request->all());
-        if (!$result) return response()->json('fail');
-
-        return response()->json('success');
-    }
-
-
-    public function apiDestroy()
-    {
-        $access = Auth::user()->hasAccess('CATEGORY_DELETE');
-        if (!$access) return response()->json('No access');
-
-        $result = $this->item->delete();
-        if (!$result) return response()->json('fail');
 
         return response()->json(
             [
@@ -99,21 +77,31 @@ class PageCategoryEntity implements WebEntity
     }
 
 
-    static function mass($request)
-    {
-        $data = $request->all();
+    // public function apiUpdate(Request $request)
+    // {
+    //     $access = Auth::user()->hasAccess('CATEGORY_EDIT');
+    //     if (!$access) return response()->json('No access');
 
-        if (empty($data['mass_edit'])) {
-            return redirect()->back()->with('error', __('admin/messages.categories.mass.errors.no_categories'));
-        }
+    //     $result = $this->item->update($request->all());
+    //     if (!$result) return response()->json('fail');
 
-        switch ($data['mass_action']) {
-            case 'delete':
-                Auth::user()->hasAccessOrRedirect('CATEGORY_DELETE');
-                PageCategory::whereIn('id', $data['mass_edit'])->delete();
-                break;
-        }
+    //     return response()->json('success');
+    // }
 
-        return redirect(route('admin.pages.categories.index'));
-    }
+
+    // public function apiDestroy()
+    // {
+    //     $access = Auth::user()->hasAccess('CATEGORY_DELETE');
+    //     if (!$access) return response()->json('No access');
+
+    //     $this->item->delete();
+
+    //     return response()->json(
+    //         [
+    //             'message' => __('admin/messages.categories.delete.success'),
+    //             'id' => $this->item->id
+    //         ],
+    //         200
+    //     );
+    // }
 }
